% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{alpha_mul}
\alias{alpha_mul}
\title{Multiply alpha channel by given value}
\usage{
alpha_mul(x, alpha = 1)
}
\arguments{
\item{x}{hex colour with alpha with total of 8 hex digits}

\item{alpha}{alpha [0, 1]}
}
\description{
Multiply alpha channel by given value
}
