% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{alpha_set}
\alias{alpha_set}
\title{Set alpha channel}
\usage{
alpha_set(x, alpha = 1)
}
\arguments{
\item{x}{hex colour with alpha with total of 8 hex digits}

\item{alpha}{alpha [0, 1]}
}
\description{
Set alpha channel
}
