% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-css.R
\name{cat_indent}
\alias{cat_indent}
\title{A version of cat which has an indentation depth}
\usage{
cat_indent(depth, ..., sep = "")
}
\arguments{
\item{depth}{the indentation depth}

\item{...}{arguments passed to cat()}

\item{sep}{argument passed to cat()}
}
\description{
A version of cat which has an indentation depth
}
