% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-style.R
\name{css_apply}
\alias{css_apply}
\title{Create a named list (indexed by \code{xpath}) of the final computed style for each element}
\usage{
css_apply(xml, css = list(), svg = FALSE)
}
\arguments{
\item{xml}{xml document as returned by \code{xml2::read_html()} or
\code{xml2::read_xml()}}

\item{css}{list of rules parsed from CSS stylesheet. E.g as returned
by \code{cssparser::read_css()}}

\item{svg}{include SVG presentation tags in the cascade? default: FALSE}
}
\value{
named list of styles where the name is the \code{xpath} to a
        node in the document, and the value is a named list of
        property/values for this element.
}
\description{
Create a named list (indexed by \code{xpath}) of the final computed style for each element
}
