% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-inline-style.R
\name{css_apply_inline}
\alias{css_apply_inline}
\title{Apply the CSS to the given HTML, storing the result as inline 'style' tags on each element}
\usage{
css_apply_inline(xml, css)
}
\arguments{
\item{xml}{html/xml document (as read by \code{xml2::read_xml()})}

\item{css}{CSS stylesheet to apply  (as read by \code{cssparser::read_css()})}
}
\value{
new xml document with final computed CSS written to inline style
        attribute on each element.
}
\description{
Apply the CSS to the given HTML, storing the result as inline 'style' tags on each element
}
