% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{css_colour_to_hex}
\alias{css_colour_to_hex}
\title{Convert CSS colour to standard hex colour (with alpha0)}
\usage{
css_colour_to_hex(x)
}
\arguments{
\item{x}{CSS colour}
}
\value{
hex colour (always with alpha channel)
}
\description{
CSS colours can be lots of things:
}
\details{
\itemize{
\item{Hex colour with 3, 6 or 8 characters}
\item{CSS colour name e.g. 'red'.  Not all CSS colours correspond to
      R colours e.g. 'silver' is not in R}
\item{rgb(), hsl(), hcl(), lab() and other colourspace-specific colour constructores.
      Only rgb and hsl are currently handled}
\item{color() for complex colourspace specifidation. Not handled yet}
\item{'transparent' and sometimes 'none' to indicate '#00000000'}
\item{'currentColor' or 'currentcolor' to indicate the colour for this element
      should be taken from whatever the current 'color' property is}
\item{for SVG, colours colours can be references to 'linearGradient' and 'radialGradient'
      specifications, or patterns}
}

Ref: \url{https://developer.mozilla.org/en-US/docs/Web/CSS/color_value#currentcolor_keyword}
}
