% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric-values.R
\name{css_length}
\alias{css_length}
\title{Create a new CSS length object}
\usage{
css_length(x, unit)
}
\arguments{
\item{x}{numeric value}

\item{unit}{character string for unit}
}
\description{
Create a new CSS length object
}
