% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric-values.R
\name{css_length_as_pixels}
\alias{css_length_as_pixels}
\title{Convert a CSS unit into a rough pixel measurement}
\usage{
css_length_as_pixels(x, font_size = 12, ...)
}
\arguments{
\item{x}{object of type \code{css_length} as returned by \code{css_string_to_css_length()}}

\item{font_size}{font size to calculate 'em' with. default: 12}

\item{...}{other arguments ignored}
}
\value{
simple numeric value
}
\description{
This is a naive conversion.
}
\details{
For a proper conversion, it would have to take into account the font-size
on the root element (for \code{rem} units), and various things to do with
framepoint and viewing size of the element and its parents.
}
