% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-styles.R
\name{css_merge}
\alias{css_merge}
\title{Merge multiple stylesheets given in priority order}
\usage{
css_merge(...)
}
\arguments{
\item{...}{multiple
CSS objects. The order in which these arguments are given reflect the
priority of the stylesheets from lowest to highest priority.  Later stylesheets
(i.e. high priority) will override any styles declared earlier
(lower priority)}
}
\value{
final cascaded stylesheet
}
\description{
Merge multiple stylesheets given in priority order
}
\examples{
\dontrun{
css1 <- read_css("chrome_builtin.css")
css2 <- read_css("this_page.css")
css_merge(css1, css2)
}

}
