% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-styles.R
\name{css_merge_core}
\alias{css_merge_core}
\title{A CSS-aware version of nested \code{modifyList()}}
\usage{
css_merge_core(base, new)
}
\arguments{
\item{base}{the original list}

\item{new}{the new list from which to take values and put into base}
}
\value{
updated version of base list
}
\description{
If an individual element in the base list has the 'important' attribute
set to TRUE, then it does *not* get overwritten.
}
\details{
Some properties are accumulative, rather than replacement-based.  E.g.
SVGs \code{transform} attribute combines parent + child transform
matrices.  For now, any accumulative properties are concatenated into
a vector of character stirngs to be dealt with by the user after
parsing.

Otherwise, this function behaves very much like \code{utils::modifyList()}
}
