% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-css.R
\name{css_pretty_print}
\alias{css_pretty_print}
\title{Pretty Printing of a CSS stylesheet (as produced by \code{read_css()})}
\usage{
css_pretty_print(x, depth = 0, rulesep = "\\n", ...)
}
\arguments{
\item{x}{object representing a CSS stylesheet}

\item{depth}{the recursion depth of this block (used for indentation)}

\item{rulesep}{separator between multiple rules}

\item{...}{other arguments ignored}
}
\description{
Pretty Printing of a CSS stylesheet (as produced by \code{read_css()})
}
