% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric-values.R
\name{css_string_as_css_length}
\alias{css_string_as_css_length}
\title{Convert a string value into a numeric value with a 'unit' attribute (if a unit is present)}
\usage{
css_string_as_css_length(x)
}
\arguments{
\item{x}{a value from CSS.}
}
\value{
\code{css_length} object
}
\description{
\url{https://www.w3schools.com/CSSref/css_units.asp}
}
