% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-agent-style-sheets.R
\docType{data}
\name{html4_user_agent_css}
\alias{html4_user_agent_css}
\title{HTML4 user-agent css}
\format{
An object of class \code{list} of length 120.
}
\usage{
html4_user_agent_css
}
\description{
HTML4 user-agent css
}
\keyword{datasets}
