% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stream.R
\name{init_stream}
\alias{init_stream}
\title{A minimal stream object (reminiscent of an R6 class)}
\usage{
init_stream(tokens)
}
\arguments{
\item{tokens}{named}
}
\description{
This tream object (an environment) is used to encapsulate a
vector of named tokens, an index of the current position, and some simple
tools for advancing the stream, consuming tokens, etc
}
