% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_char1}
\alias{is_char1}
\title{Test if string is a single string with something in it}
\usage{
is_char1(x)
}
\arguments{
\item{x}{string}
}
\value{
TRUE if string is single string with something in it
}
\description{
Test if string is a single string with something in it
}
