% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-css.R
\name{parse_at}
\alias{parse_at}
\title{Parse a simple "at' rule from the stream at the current location}
\usage{
parse_at(stream)
}
\arguments{
\item{stream}{stream}
}
\description{
This is really generic and lumps everything up to the next semi-colon into
a single character string.
}
