% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour-hex.R
\name{parse_colour_hex}
\alias{parse_colour_hex}
\title{Normalize a hex colour to an 8-char hex}
\usage{
parse_colour_hex(hex_colour)
}
\arguments{
\item{hex_colour}{e.g. '#000'}
}
\value{
8 char hex colour
}
\description{
\itemize{
\item{If hex colour doesn't have an alpha channel, add it as 'FF'}
\item{If hex colour only has 3 digits, expand it to 6, then add alpha channel}
}
}
\examples{
\dontrun{
parse_colour_hex('#123') # -> '#112233FF'
}
}
