% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour-hsl.R
\name{parse_colour_hsl_to_hex}
\alias{parse_colour_hsl_to_hex}
\title{Parse a hsl/hsla colour spec to a hex colour}
\usage{
parse_colour_hsl_to_hex(colour_func_text)
}
\arguments{
\item{colour_func_text}{string representing an \code{hsl()} or \code{hsla()} colour
e.g. \code{'hsl(50, 0.5, 0.5)'}}
}
\value{
hex colour
}
\description{
MDN \url{https://developer.mozilla.org/en-US/docs/Web/CSS/color_value/hsl()}
}
\examples{
\dontrun{
parse_colour_hsl_to_hex('hsl(50, 0.5, 0.5)')
}
}
