% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour-rgb.R
\name{parse_colour_rgb_to_hex}
\alias{parse_colour_rgb_to_hex}
\title{Parse an rgb() or rgba() colour spec to a hex colour}
\usage{
parse_colour_rgb_to_hex(rgb_text)
}
\arguments{
\item{rgb_text}{string representing an \code{rgb()} or \code{rgba()} colour
e.g. \code{'rgb(255, 0, 0)'}}
}
\value{
hex colour
}
\description{
MDN \url{https://developer.mozilla.org/en-US/docs/Web/CSS/color_value/rgb()}
}
\examples{
\dontrun{
parse_colour_rgb_to_hex('rgb(255, 0, 0)')
}
}
