% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-css.R
\name{parse_declaration}
\alias{parse_declaration}
\title{Extract a declaration (i.e. a property/value pair) from token stream}
\usage{
parse_declaration(stream)
}
\arguments{
\item{stream}{environment containing tokens and the current index. This is
an internal only datastructure to assist in parsing}
}
\value{
a named list i.e. the name/value property pair
}
\description{
e.g. "fill: black; color: blue" -> list(fill = 'black', color = 'blue')
}
