% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-css.R
\name{parse_rules_from_stream}
\alias{parse_rules_from_stream}
\title{Parse rules from the stream at the current location}
\usage{
parse_rules_from_stream(stream)
}
\arguments{
\item{stream}{token stream}
}
\value{
name list of list(rule_name = list(property = value, ...))
}
\description{
A 'rule' is a 'selector' + a list of property/value pairs.
}
