% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stream.R
\name{print.stream}
\alias{print.stream}
\title{s3 method fr printing a stream}
\usage{
\method{print}{stream}(x, ...)
}
\arguments{
\item{x}{stream object}

\item{...}{other arguments passed to print()}
}
\description{
s3 method fr printing a stream
}
