% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-css.R
\name{read_css}
\alias{read_css}
\title{Convert CSS stylesheet from text or a file into a list object}
\usage{
read_css(stylesheet)
}
\arguments{
\item{stylesheet}{CSS stylsheet. Either as a character string containing the CSS,
or a path to a file containing CSS text}
}
\value{
named list of lists, where the top-level name is the CSS selector,
        and the value is a list of property/value pairs for this selector
        (i.s. a CSS declaration block)
}
\description{
Convert CSS stylesheet from text or a file into a list object
}
