% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-inline-style.R
\name{read_inline_style}
\alias{read_inline_style}
\title{Parse CSS declaration block (aka inline style) to a named list of CSS declarations}
\usage{
read_inline_style(inline_style)
}
\arguments{
\item{inline_style}{set of ";"-delmited declarations i.e. an inline style,
or a CSS declaration block: e.g.
"stroke:red; fill: black; stroke-width:12"}
}
\value{
named list of declarations (i.e. property/value pairs)
        e.g. \code{list(stroke = 'red', fill = 'black', 'stroke-width' = 12)}
}
\description{
Parse CSS declaration block (aka inline style) to a named list of CSS declarations
}
