% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-style.R
\name{recurse_node_tree}
\alias{recurse_node_tree}
\title{Depth-first recursion into the HTML document node tree to accumulate/cascade styles}
\usage{
recurse_node_tree(node, xpath_styles, parent_style = list(), svg)
}
\arguments{
\item{node}{current XML node}

\item{xpath_styles}{list containing the accumulated styles for each xpath}

\item{parent_style}{the style from the direct parent}

\item{svg}{include SVG presentation tags in the cascade? default: FALSE}
}
\value{
updated xpath style lookup list
}
\description{
Depth-first recursion into the HTML document node tree to accumulate/cascade styles
}
