% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selector-specificity.R
\name{selector_specificity}
\alias{selector_specificity}
\title{Calculate selector specificity}
\usage{
selector_specificity(sel)
}
\arguments{
\item{sel}{selector}
}
\value{
numeric value
}
\description{
Algorithm defined here: \url{https://drafts.csswg.org/selectors-3/#specificity}
}
\details{
A selector's specificity is calculated as follows:
\itemize{
  \item{count the number of ID selectors in the selector (= a)}
  \item{count the number of class selectors, attributes selectors, and
         pseudo-classes in the selector (= b)}
  \item{count the number of type selectors and pseudo-elements in the selector (= c)}
  \item{ignore the universal selector}
}

Selectors inside the negation pseudo-class are counted like any other,
but the negation itself does not count as a pseudo-class.

Concatenating the three numbers a-b-c (in a number system with a large base)
gives the specificity.

Since we only really have base-10 numbers (rather than numbers in
arbitrarily large bases), just use 2 digits for each of the counts.  Hopefully it
is really unlikely that any of the counts exceeds 99!
}
