% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{split_font_shorthand}
\alias{split_font_shorthand}
\title{Split font shorthand property into a named list of verbose "font-*" properties}
\usage{
split_font_shorthand(font_text)
}
\arguments{
\item{font_text}{the text of the "font" declaration e.g.
"italic 1.2em "Fira Sans", serif;"}
}
\value{
list of font-related property/value declarations e.g.
        \code{list('font-family'='serif', font-size='1.2em', ...)}
}
\description{
This is not perfect/finished but it is a step in the right direction.
}
\details{
For example: this function does not split the "font-variant" shorthand
}
