% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply-inline-style.R
\name{style_flatten_to_inline}
\alias{style_flatten_to_inline}
\title{Flatten a style to an inline string}
\usage{
style_flatten_to_inline(style)
}
\arguments{
\item{style}{a named list of property/value pairs}
}
\value{
single string suitable for a \code{style} attribute on an element
}
\description{
Flatten a style to an inline string
}
\examples{
\dontrun{
style_flatten_to_line(list(color='black', border='1px')) # -> "color:black; border: 1px;"
}
}
