% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-styles.R
\name{style_merge}
\alias{style_merge}
\title{Merge multiple styles for a given element which are given in priority order}
\usage{
style_merge(...)
}
\arguments{
\item{...}{multiple style objects (i.e. declaration blocks).
The order in which these arguments are given reflect the
priority of the styles from lowest to highest priority.  Later styles
(i.e. high priority) will override any styles declared earlier
(lower priority)}
}
\value{
final cascaded result
}
\description{
Merge multiple styles for a given element which are given in priority order
}
\examples{
\dontrun{
style1 <- list(color = 'blue', `font-weight` = 'bold')
style2 <- list(color = 'red', `font-size`="12px")
style_merge(style1, style2)
}


}
