% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-css.R
\name{style_pretty_print}
\alias{style_pretty_print}
\title{Pretty Printing of a single style or declaration block}
\usage{
style_pretty_print(x, depth = 0)
}
\arguments{
\item{x}{object representing a list of property/value declarations}

\item{depth}{the recursion depth of this block (used for indentation)}
}
\description{
Pretty Printing of a single style or declaration block
}
