% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-xml.R
\name{xml_duplicate}
\alias{xml_duplicate}
\title{Make a distinct, unlinked, independent copy of an xml node}
\usage{
xml_duplicate(x)
}
\arguments{
\item{x}{xml document}
}
\value{
new, unlinked xml document
}
\description{
Since XML documents and nodes are almost always handled 'by-reference'
if you want to keep an original untouched when adding nodes etc,
you need to make an unlinked copy  of it.
}
